Add files via upload
This is actually a Number Guessing game that i have made. Try it once guys  
NOTE : ITS NOT A GUI GAME , its just a CLI(command line interface game) 

STEPS TO PLAY THE GAME : 
step1 : copy the code and then paste it in the online python interpreter and then run the code , and follow the steps, BOOM  thats it. 
step2 : Downlaod the current file first and then upload this file to the online compiler and then run in . You can use ONLINE GDB for doing that, or you can use any interpreter if you know some.